ref: Move getFormattedTimestamp until utils/date/
