+++ static/app/utils/date/getFormattedTimestamp.tsx
